cambié el tipo de dato de factorial

+++ Guía4/eaprox.hs
@@ -2,8 +2,8 @@
 a partir de la siguiente sumatoria:
 eˆ(n) = sumatoria i=0 n  de 1/i!
 -}
-
-factorial :: Integer -> Integer
+-- Intento hacer la recursión en 
+factorial :: Integer -> Float
 factorial x | x == 0 = 1
             | otherwise = x * factorial (x-1)
